Use action filters to check if user can do things
Add an attribute to actions which can only be performed by a user who is a participant in the game. This makes the tests icky though, as the filter is not applied automatically to the mocks, and has to be called by hand, which isn't testing that the attribute has been placed on the right methods.