Rudimentary draw mechanism
A rejection does not notify the other player.
A player can offer a draw at any time (which is against the rules).
The fact that an offer has been made is not persisted in the database.
No check is made that the player posting AcceptDraw is the player to whom the offer was made.