The padlock icon does not scale correctly.
The old behavior was to have it as 128x128 at all times, but now the board scales, the size of a 2x2 square might be different (2*68, for instance, instead of 2*64). This CSS doesn't quite work (for some sizes, it tiles the background image, even though background-repeat is set to no-repeat), but it's better than the old behavior.