Make the board remember the moves after ajax
Because the board redraws after being resized, it needs to re-highlight the last move played. Previously this was only available in the model when the page was originally rendered on the server, but now the signalR callback will save the last move in the game object so that it can be re-rendered.